requirements.txt
django==2.2.28
dropbox==8.7.1
gunicorn==19.7.1
djangorestframework==3.11.2
django-filter==2.4.0
coreapi==2.3.3
django-cors-headers==2.2.0
simplejson==3.15.0
djangorestframework-datatables==0.3.0
gmplot==1.2.0
drf-nested-routers==0.90.2
django-extensions==2.1.6
firebase-admin==3.2.1
